feat: make marketplace reusable section
